chore: refactor startup logic to work without reload (#3146)
* refactor workspace activation logic

* reload ws

* fix merge issues

* refactor workspace activation logic

* snap

* step1

* init vaults

* check for dup vaults

* simplify

* update tests

* update

* update diagnostics

* update diag report

* update

* fix bugs + tree view

* move reload logic out

* update trait nit

* allow for launching without reload

* more refactoring

* update

* remove dangling anchor

* pr changes
