+++ packages/plugin-core/src/_extension.ts
@@ -1,19 +1,14 @@
-import { SubProcessExitType } from "@dendronhq/api-server";
 import {
   CONSTANTS,
-  DendronError,
+  DWorkspaceV2,
   getStage,
-  GitEvents,
   GraphEvents,
   GraphThemeEnum,
   GraphThemeTestGroups,
   GRAPH_THEME_TEST,
   InstallStatus,
   isDisposable,
-  TreeViewItemLabelTypeEnum,
-  VaultUtils,
   VSCodeEvents,
-  WorkspaceType,
 } from "@dendronhq/common-all";
 import {
   getDurationMilliseconds,
@@ -24,12 +19,10 @@ import {
 import {
   HistoryService,
   MetadataService,
-  WorkspaceService,
   WorkspaceUtils,
 } from "@dendronhq/engine-server";
 import * as Sentry from "@sentry/node";
 import fs from "fs-extra";
-import _ from "lodash";
 import os from "os";
 import path from "path";
 import semver from "semver";
@@ -62,24 +55,20 @@ import ReferenceProvider from "./features/ReferenceProvider";
 import RenameProvider from "./features/RenameProvider";
 import { KeybindingUtils } from "./KeybindingUtils";
 import { Logger } from "./logger";
-import { EngineAPIService } from "./services/EngineAPIService";
 import { StateService } from "./services/stateService";
-import { TextDocumentServiceFactory } from "./services/TextDocumentServiceFactory";
 import { Extensions } from "./settings";
 import { FeatureShowcaseToaster } from "./showcase/FeatureShowcaseToaster";
 import { AnalyticsUtils, sentryReportingCallback } from "./utils/analytics";
 import { isAutoCompletable } from "./utils/AutoCompletable";
 import { ExtensionUtils } from "./utils/ExtensionUtils";
 import { AutoCompletableRegistrar } from "./utils/registers/AutoCompletableRegistrar";
+import { StartupPrompts } from "./utils/StartupPrompts";
 import { StartupUtils } from "./utils/StartupUtils";
-import { EngineNoteProvider } from "./views/EngineNoteProvider";
-import { NativeTreeView } from "./views/NativeTreeView";
 import { VSCodeUtils } from "./vsCodeUtils";
 import { showWelcome } from "./WelcomeUtils";
 import { DendronExtension, getDWorkspace, getExtension } from "./workspace";
 import { TutorialInitializer } from "./workspace/tutorialInitializer";
-import { WorkspaceActivator } from "./workspace/workspaceActivater";
-import { WorkspaceInitFactory } from "./workspace/WorkspaceInitFactory";
+import { WorkspaceActivator } from "./workspace/workspaceActivator";
 import { WSUtils } from "./WSUtils";
 
 const MARKDOWN_WORD_PATTERN = new RegExp("([\\w\\.\\#]+)");
@@ -108,97 +97,6 @@ export function activate(
     });
   }
   return context;
-}
-
-async function reloadWorkspace(opts: { extension: DendronExtension }) {
-  const ctx = "reloadWorkspace";
-  const { extension } = opts;
-  const ws = extension.getDWorkspace();
-  const maybeEngine = await WSUtils.reloadWorkspace();
-  if (!maybeEngine) {
-    return maybeEngine;
-  }
-  Logger.info({ ctx, msg: "post-ws.reloadWorkspace" });
-
-  // Run any initialization code necessary for this workspace invocation.
-  const initializer = WorkspaceInitFactory.create();
-
-  if (initializer?.onWorkspaceOpen) {
-    initializer.onWorkspaceOpen({ ws });
-  }
-
-  vscode.window.showInformationMessage("Dendron is active");
-  Logger.info({ ctx, msg: "exit" });
-
-  await postReloadWorkspace({ extension });
-  HistoryService.instance().add({
-    source: "extension",
-    action: "initialized",
-  });
-  return maybeEngine;
-}
-
-async function postReloadWorkspace(opts: { extension: DendronExtension }) {
-  const ctx = "postReloadWorkspace";
-  const { extension } = opts;
-  const wsService = extension.workspaceService;
-  if (!wsService) {
-    const errorMsg = "No workspace service found.";
-    Logger.error({
-      msg: errorMsg,
-      error: new DendronError({ message: errorMsg }),
-    });
-    return;
-  }
-
-  const wsMeta = wsService.getMeta();
-  const previousWsVersion = wsMeta.version;
-  // stats
-  // NOTE: this is legacy to upgrade .code-workspace specific settings
-  // we are moving everything to dendron.yml
-  // see [[2021 06 Deprecate Workspace Settings|proj.2021-06-deprecate-workspace-settings]]
-  if (previousWsVersion === CONSTANTS.DENDRON_INIT_VERSION) {
-    Logger.info({ ctx, msg: "no previous global version" });
-    vscode.commands
-      .executeCommand(DENDRON_COMMANDS.UPGRADE_SETTINGS.key)
-      .then((changes) => {
-        Logger.info({ ctx, msg: "postUpgrade: new wsVersion", changes });
-      });
-    wsService.writeMeta({ version: DendronExtension.version() });
-  } else {
-    const newVersion = DendronExtension.version();
-    if (semver.lt(previousWsVersion, newVersion)) {
-      let changes: any;
-      Logger.info({ ctx, msg: "preUpgrade: new wsVersion" });
-      try {
-        changes = await vscode.commands.executeCommand(
-          DENDRON_COMMANDS.UPGRADE_SETTINGS.key
-        );
-        Logger.info({
-          ctx,
-          msg: "postUpgrade: new wsVersion",
-          changes,
-          previousWsVersion,
-          newVersion,
-        });
-        wsService.writeMeta({ version: DendronExtension.version() });
-      } catch (err) {
-        Logger.error({
-          msg: "error upgrading",
-          error: new DendronError({ message: JSON.stringify(err) }),
-        });
-        return;
-      }
-      HistoryService.instance().add({
-        source: "extension",
-        action: "upgraded",
-        data: { changes },
-      });
-    } else {
-      Logger.info({ ctx, msg: "same wsVersion" });
-    }
-  }
-  Logger.info({ ctx, msg: "exit" });
 }
 
 // Only exported for test purposes ^jtm6bf7utsxy
@@ -281,9 +179,27 @@ export async function _activate(
     const ws = await DendronExtension.getOrCreate(context, {
       skipSetup: stage === "test",
     });
+    const existingCommands = await vscode.commands.getCommands();
 
     // Setup the commands
     await _setupCommands({ ext: ws, context, requireActiveWorkspace: false });
+    // Order matters. Need to register `Reload Index` command before activating workspace
+    // Workspace activation calls `RELOAD_INDEX` via {@link WSUtils.reloadWorkspace}
+    if (!existingCommands.includes(DENDRON_COMMANDS.RELOAD_INDEX.key)) {
+      context.subscriptions.push(
+        vscode.commands.registerCommand(
+          DENDRON_COMMANDS.RELOAD_INDEX.key,
+          sentryReportingCallback(async (silent?: boolean) => {
+            const out = await new ReloadIndexCommand().run({ silent });
+            if (!silent) {
+              vscode.window.showInformationMessage(`finish reload`);
+            }
+            return out;
+          })
+        )
+      );
+    }
+    await _setupCommands({ ext: ws, context, requireActiveWorkspace: true });
 
     if (!opts?.skipLanguageFeatures) {
       _setupLanguageFeatures(context);
@@ -381,65 +297,21 @@ export async function _activate(
       if (!maybeWsRoot) {
         return false;
       }
-      const maybeWs = await activator.activate({
+      const resp = await activator.init({
         ext: ws,
         context,
         wsRoot: maybeWsRoot,
       });
-      if (!maybeWs) {
+      if (resp.error) {
         return false;
       }
-      const wsImpl = maybeWs;
-      const start = process.hrtime();
-      const dendronConfig = wsImpl.config;
-
-      // --- Get Version State
-      const wsRoot = wsImpl.wsRoot;
-      const wsService = new WorkspaceService({ wsRoot });
-      let previousWorkspaceVersionFromWSService = wsService.getMeta().version;
-
-      // Fix a temporary issue where CLI was writing an invalid version number
-      // to .dendron.ws:
-      if (previousWorkspaceVersionFromWSService === "dendron-cli") {
-        previousWorkspaceVersionFromWSService = "0.91.0";
-      }
-
-      if (
-        !semver.valid(previousWorkspaceVersionFromWSService) ||
-        semver.gt(
-          previousWorkspaceVersionFromState,
-          previousWorkspaceVersionFromWSService
-        )
-      ) {
-        previousWorkspaceVersionFromWSService =
-          previousWorkspaceVersionFromState;
-        wsService.writeMeta({ version: previousGlobalVersionFromState });
-      }
-      const previousWorkspaceVersion = previousWorkspaceVersionFromWSService;
+      const wsImpl: DWorkspaceV2 = resp.data.workspace;
 
       // initialize Segment client
       AnalyticsUtils.setupSegmentWithCacheFlush({ context, ws: wsImpl });
 
-      const maybeWsSettings =
-        ws.type === WorkspaceType.CODE
-          ? wsService.getCodeWorkspaceSettingsSync()
-          : undefined;
-      if (!opts?.skipMigrations) {
-        await StartupUtils.runMigrationsIfNecessary({
-          wsService,
-          currentVersion,
-          previousWorkspaceVersion,
-          maybeWsSettings,
-          dendronConfig,
-        });
-      }
-
-      if (opts?.skipInteractiveElements) {
-        // check for duplicate config keys and prompt for a fix.
-        StartupUtils.showDuplicateConfigEntryMessageIfNecessary({
-          ext: ws,
-        });
-
+      // show interactive elements when **extension starts**
+      if (!opts?.skipInteractiveElements) {
         // check for missing default config keys and prompt for a backfill.
         StartupUtils.showMissingDefaultConfigMessageIfNecessary({
           ext: ws,
@@ -456,50 +328,6 @@ export async function _activate(
       // Re-use the id for error reporting too:
       Sentry.setUser({ id: SegmentClient.instance().anonymousId });
 
-      const didClone = await wsService.initialize({
-        onSyncVaultsProgress: () => {
-          vscode.window.showInformationMessage(
-            "found empty remote vaults that need initializing"
-          );
-        },
-        onSyncVaultsEnd: () => {
-          vscode.window.showInformationMessage(
-            "finish initializing remote vaults. reloading workspace"
-          );
-          setTimeout(VSCodeUtils.reloadWindow, 200);
-        },
-      });
-      if (didClone) {
-        return false;
-      }
-
-      ws.workspaceService = wsService;
-      // set vaults now that ws is initialized
-      const vaults = wsService.vaults;
-
-      // check for vaults with same name
-      const uniqVaults = _.uniqBy(vaults, (vault) => VaultUtils.getName(vault));
-      if (_.size(uniqVaults) < _.size(vaults)) {
-        const txt = "Fix it";
-        await vscode.window
-          .showErrorMessage(
-            "Multiple Vaults with the same name. See https://dendron.so/notes/a6c03f9b-8959-4d67-8394-4d204ab69bfe.html#multiple-vaults-with-the-same-name to fix",
-            txt
-          )
-          .then((resp) => {
-            if (resp === txt) {
-              vscode.commands.executeCommand(
-                "vscode.open",
-                vscode.Uri.parse(
-                  "https://dendron.so/notes/a6c03f9b-8959-4d67-8394-4d204ab69bfe.html#multiple-vaults-with-the-same-name"
-                )
-              );
-            }
-          });
-        return false;
-      }
-      wsService.writeMeta({ version: DendronExtension.version() });
-
       // stats
       const platform = getOS();
       const extensions = Extensions.getDendronExtensionRecommendations().map(
@@ -513,113 +341,24 @@ export async function _activate(
       );
 
       Logger.info({
-        ctx,
+        ctx: ctx + ":postSetupWorkspace",
         platform,
         extensions,
         vaults: wsImpl.vaults,
       });
 
       // --- Start Initializating the Engine
-      WSUtils.showInitProgress();
-
-      const port = await ExtensionUtils.startServerProcess({
-        context,
-        start,
-        wsService,
-        onExit: (type: SubProcessExitType) => {
-          const txt = "Restart Dendron";
-          vscode.window
-            .showErrorMessage("Dendron engine encountered an error", txt)
-            .then(async (resp) => {
-              if (resp === txt) {
-                AnalyticsUtils.track(VSCodeEvents.ServerCrashed, {
-                  code: type,
-                });
-                _activate(context);
-              }
-            });
-        },
-      });
-
-      // Setup the Engine API Service and the tree view
-      const engineAPIService = updateEngineAPI(port, ws);
-
-      // TODO: This should eventually be consolidated with other view setup
-      // logic as in workspace.ts, but right now this needs an instance of
-      // EngineAPIService for init
-
-      const providerConstructor = function () {
-        return new EngineNoteProvider(engineAPIService);
-      };
-
-      const existingCommands = await vscode.commands.getCommands();
-      if (opts?.skipTreeView) {
-        const treeView = new NativeTreeView(providerConstructor);
-        treeView.show();
-        _setupTreeViewCommands(treeView, existingCommands);
-        context.subscriptions.push(treeView);
-      }
-
-      // Instantiate TextDocumentService
-      context.subscriptions.push(TextDocumentServiceFactory.create(ws));
-
-      // Order matters. Need to register `Reload Index` command before `reloadWorkspace`
-      if (!existingCommands.includes(DENDRON_COMMANDS.RELOAD_INDEX.key)) {
-        context.subscriptions.push(
-          vscode.commands.registerCommand(
-            DENDRON_COMMANDS.RELOAD_INDEX.key,
-            sentryReportingCallback(async (silent?: boolean) => {
-              const out = await new ReloadIndexCommand().run({ silent });
-              if (!silent) {
-                vscode.window.showInformationMessage(`finish reload`);
-              }
-              return out;
-            })
-          )
-        );
-      }
-      const reloadSuccess = await reloadWorkspace({ extension: ws });
-      const durationReloadWorkspace = getDurationMilliseconds(start);
-
-      await ExtensionUtils.trackWorkspaceInit({
-        durationReloadWorkspace,
-        activatedSuccess: !!reloadSuccess,
+      const wsService = resp.data.wsService;
+      const respActivate = await activator.activate({
         ext: ws,
+        context,
+        wsRoot: maybeWsRoot,
+        engine: resp.data.engine,
+        wsService,
       });
-
-      if (!reloadSuccess) {
-        HistoryService.instance().add({
-          source: "extension",
-          action: "not_initialized",
-        });
+      if (respActivate.error) {
         return false;
       }
-
-      ExtensionUtils.setWorkspaceContextOnActivate(wsService.config);
-
-      MetadataService.instance().setDendronWorkspaceActivated();
-      await _setupCommands({ ext: ws, context, requireActiveWorkspace: true });
-
-      // Track contributors to repositories, but do so in the background so
-      // initialization isn't delayed.
-      const startGetAllReposNumContributors = process.hrtime();
-      wsService
-        .getAllReposNumContributors()
-        .then((numContributors) => {
-          AnalyticsUtils.track(GitEvents.ContributorsFound, {
-            maxNumContributors: _.max(numContributors),
-            duration: getDurationMilliseconds(startGetAllReposNumContributors),
-          });
-        })
-        .catch((err) => {
-          Sentry.captureException(err);
-        });
-
-      if (stage !== "test") {
-        await ws.activateWatchers();
-        togglePluginActiveContext(true);
-      }
-
       if (!opts?.skipInteractiveElements) {
         // on first install, warn if extensions are incompatible ^dlx35gstwsun
         if (extensionInstallStatus === InstallStatus.INITIAL_INSTALL) {
@@ -632,17 +371,11 @@ export async function _activate(
           showcase.showToast();
         }, ONE_MINUTE_IN_MS);
       }
-
-      // Add the current workspace to the recent workspace list. The current
-      // workspace is either the workspace file (Code Workspace) or the current
-      // folder (Native Workspace)
-      const workspace = DendronExtension.tryWorkspaceFile()?.fsPath || wsRoot;
-      MetadataService.instance().addToRecentWorkspaces(workspace);
-
-      Logger.info({ ctx, msg: "fin startClient", durationReloadWorkspace });
     } else {
       // ws not active
       Logger.info({ ctx, msg: "dendron not active" });
+      AnalyticsUtils.setupSegmentWithCacheFlush({ context });
+      Sentry.setUser({ id: SegmentClient.instance().anonymousId });
     }
 
     if (extensionInstallStatus === InstallStatus.INITIAL_INSTALL) {
@@ -794,7 +527,7 @@ async function showWelcomeOrWhatsNew({
 
   // Show lapsed users (users who have installed Dendron but haven't initialied
   // a workspace) a reminder prompt to re-engage them.
-  StartupUtils.showLapsedUserMessageIfNecessary({ assetUri });
+  StartupPrompts.showLapsedUserMessageIfNecessary({ assetUri });
 
   // Show inactive users (users who were active on first week but have not used lookup in 2 weeks)
   // a reminder prompt to re-engage them.
@@ -995,70 +728,4 @@ function _setupLanguageFeatures(context: vscode.ExtensionContext) {
   );
   completionProvider.activate(context);
   codeActionProvider.activate(context);
-}
-
-// ^qxkkg70u6w0z
-
-function _setupTreeViewCommands(
-  treeView: NativeTreeView,
-  existingCommands: string[]
-) {
-  if (
-    !existingCommands.includes(DENDRON_COMMANDS.TREEVIEW_LABEL_BY_TITLE.key)
-  ) {
-    vscode.commands.registerCommand(
-      DENDRON_COMMANDS.TREEVIEW_LABEL_BY_TITLE.key,
-      sentryReportingCallback(() => {
-        treeView.updateLabelType({
-          labelType: TreeViewItemLabelTypeEnum.title,
-        });
-      })
-    );
-  }
-
-  if (
-    !existingCommands.includes(DENDRON_COMMANDS.TREEVIEW_LABEL_BY_FILENAME.key)
-  ) {
-    vscode.commands.registerCommand(
-      DENDRON_COMMANDS.TREEVIEW_LABEL_BY_FILENAME.key,
-      sentryReportingCallback(() => {
-        treeView.updateLabelType({
-          labelType: TreeViewItemLabelTypeEnum.filename,
-        });
-      })
-    );
-  }
-
-  /**
-   * This is a little flaky right now, but it works most of the time.
-   * Leaving this for dev / debug purposes.
-   * Enablement is set to be DendronContext.DEV_MODE
-   *
-   * TODO: fix tree item register issue and flip the dev mode flag.
-   */
-  if (!existingCommands.includes(DENDRON_COMMANDS.TREEVIEW_EXPAND_ALL.key)) {
-    vscode.commands.registerCommand(
-      DENDRON_COMMANDS.TREEVIEW_EXPAND_ALL.key,
-      sentryReportingCallback(async () => {
-        await treeView.expandAll();
-      })
-    );
-  }
-}
-
-function updateEngineAPI(
-  port: number | string,
-  ext: DendronExtension
-): EngineAPIService {
-  // set engine api ^9dr6chh7ah9v
-  const svc = EngineAPIService.createEngine({
-    port,
-    enableWorkspaceTrust: vscode.workspace.isTrusted,
-    vaults: ext.getDWorkspace().vaults,
-    wsRoot: ext.getDWorkspace().wsRoot,
-  });
-  ext.setEngine(svc);
-  ext.port = _.toInteger(port);
-
-  return svc;
 }
