+++ packages/plugin-core/src/commands/LaunchTutorialWorkspaceCommand.ts
@@ -44,7 +44,8 @@ export class LaunchTutorialWorkspaceCommand extends BasicCommand<
     await new SetupWorkspaceCommand().execute({
       rootDirRaw: filePath,
       workspaceInitializer: new TutorialInitializer(),
-      workspaceType: WorkspaceType.CODE,
+      workspaceType: WorkspaceType.NATIVE,
+      EXPERIMENTAL_openNativeWorkspaceNoReload: true,
     });
   }
 }
