+++ packages/plugin-core/src/test/suite-integ/DiagnosticsReport.test.ts
@@ -6,18 +6,16 @@ import path from "path";
 import { DiagnosticsReportCommand } from "../../commands/DiagnosticsReport";
 import { VSCodeUtils } from "../../vsCodeUtils";
 import { expect } from "../testUtilsv2";
-import { runLegacyMultiWorkspaceTest, setupBeforeAfter } from "../testUtilsV3";
+import { describeMultiWS } from "../testUtilsV3";
 
-suite("DiagnosticsReport", function () {
-  const ctx = setupBeforeAfter(this);
-
-  test("basic", (done) => {
-    runLegacyMultiWorkspaceTest({
-      ctx,
-      preSetupHook: async ({ wsRoot, vaults }) => {
-        ENGINE_HOOKS.setupBasic({ wsRoot, vaults });
-      },
-      onInit: async () => {
+suite("GIVEN DiagnosticsReport", function () {
+  describeMultiWS(
+    "WHEN run Diagnostics Report",
+    {
+      preSetupHook: ENGINE_HOOKS.setupBasic,
+    },
+    () => {
+      test("THEN generate diagnostics report", async () => {
         const logDst = path.join(
           path.dirname(env("LOG_DST")),
           "dendron.server.log"
@@ -33,8 +31,7 @@ suite("DiagnosticsReport", function () {
           match: ["foobar", "Dendron Confg", "Plugin Logs", "Workspace File"],
         });
         expect(isInString).toBeTruthy();
-        done();
-      },
-    });
-  });
+      });
+    }
+  );
 });
