+++ packages/plugin-core/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./editor";
+export * from "./StartupPrompts";
